+++ app/helpers/georgia/form_actions_helper.rb
@@ -17,6 +17,13 @@ module Georgia
       link_to "#{icon_tag('icon-retweet')} Revert".html_safe, georgia.revert_version_path(revision), method: :post, class: 'btn'
     end
 
+    def link_to_clone(model, options={})
+      return if model.nil? or model.new_record?
+      if can? :clone, model
+        link_to "#{icon_tag('icon-copy')} Copy".html_safe, url_for(controller: controller_name,id: model.id, action: :clone), options
+      end
+    end
+
     def link_to_publish(model, options={})
       return if model.nil? or model.new_record?
       if can? :publish, model
@@ -67,11 +74,16 @@ module Georgia
     def link_to_group_list(model, options={})
       html = ""
       html << content_tag('li', link_to_group_item_edit(model, options))
+      html << content_tag('li', link_to_group_item_clone(model, options)) if options[:clone]
       html << content_tag('li', link_to_group_item_publish(model, options)) if options[:publish]
       html << content_tag('li', link_to_group_item_unpublish(model, options)) if options[:publish]
       html << content_tag('li', link_to_group_item_ask_for_review(model, options)) if options[:review]
       html << content_tag('li', link_to_group_item_delete(model, options)) if options[:delete]
       content_tag 'ul', html.html_safe, class: 'dropdown-menu pull-right'
+    end
+
+    def link_to_group_item_clone(model, options={})
+      link_to "#{icon_tag('icon-copy')} Copy".html_safe, url_for(controller: controller_name, id: model.id, action: :clone) if can? :clone, model
     end
 
     def link_to_group_item_publish(model, options={})